Use Taskcluster secret as configuration source to use cache
